fix(printing): add a flush for stdout
Since I don't want to print a newline after the cube (it will push the
cube off-screen if we're at the bottom of the window already), I need to
manually flush the buffer after starting the program. This is
specifically to fix issues with resizing the window and resuming from
suspend "not printing". ReadKey will flush the buffer when in normal
operation, but if I resume after suspend/change window size while it's
waiting for a key (unlikely that this will ever not be the case),
nothing will draw until after the press of the next key.

I could turn on autoflush, but I fear that will add a performance
hit/make things look choppy. I haven't tried it though, I may at some
point if this seems insufficient.